added detection of symlinks and removed symlinking in the current directory
I think creating symlinks in the current directory is not responsibility
of the container script. The container creates environment which
resembles typical situation of a user, with data and ic provided
somewhere in REP_BASE_PROD. Then, it's up to user to possibly reset the
REP_BASE_PROD and create the links. I.e. it should be done in the
analysis script, as it is usually done anyway.